created a search box
