+++ src/index.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="/css/style.css" />
     <script src="./js/main.js" type="module"></script>
   </head>
+  
   <body>
     <header class="divider" id="main-header"></header>
     <main class="divider">
@@ -21,6 +22,7 @@
           <div>Sleep<span class="highlight">Outside</span></div>
         </div>
       </div>
+      
       <section class="mission">
         <p>We spend too much time these days indoors...sitting usually.</p>
         <p>No wonder we are a mess!</p>
@@ -35,9 +37,18 @@
         <button id="newsletter" >Get Newsletter</button>
         
       </div>
+      <div class="search-container">
+  
+        <!--<form action="/action_page.php" id="search-button">-->
+          <input type="text" placeholder="Search..." name="search" id="inputBox">
+          <button type="submit" id="searchBtn">
+            <img src="../images/search.png" alt="search">
+          </button>
+        <!--</form>-->
+      </div>
 
       <section class="products" id="products-l">
-        <h2>Top Products</h2>
+        
         <ul class="product-list">
         <div id="product-grid">
           <a href="product-list/index.html?category=tents"><img src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Tent_3544767.svg" alt="Tent Icon from Noun Project: Mustofa Bayu"></a>
