+++ src/js/productDetails.mjs
@@ -7,6 +7,8 @@ let product = {};
 export default async function productDetails(productId) {
     // get the details for the current product. findProductById will return a promise! use await or .then() to process it
     product = await findProductById(productId);
+    console.log('in product details');
+    console.log(product);
     // once we have the product details we can render out the HTML
     renderProductDetails();
     // once the HTML is rendered we can add a listener to Add to Cart button
