+++ src/js/productList.mjs
@@ -20,6 +20,6 @@ export default async function productList(selector, category) {
 
 
   renderListWithTemplate(productCardTemplate, prodList, products);
-  // document.querySelector(".title").innerHTML = category;
+  
 }
