+++ src/product-list/index.html
@@ -9,40 +9,15 @@
   </head>
   <body>
     <header class="divider" id="main-header"></header>
+    
     <main class="divider">
-      <!-- <div class="hero">
-        <img
-          src="/images/banner-sm.jpg"
-          srcset="images/banner-sm.jpg 500w, images/banner.jpg"
-          alt="image of a high mountain lake"
-        />
-        <div class="logo logo--square">
-          <img src="/images/noun_Tent_2517.svg" alt="tent image for logo" />
-          <div>Sleep<span class="highlight">Outside</span></div>
-        </div>
-      </div>
-      <section class="mission">
-        <p>We spend too much time these days indoors...sitting usually.</p>
-        <p>No wonder we are a mess!</p>
-        <p>
-          What we really need is to get outside...in fact SLEEP outside. Our
-          mission is to provide you with everything you need to do this
-          comfortably at an affordable price.
-        </p>
-      </section> -->
-<!-- 
-      <div class="mailing-list">
-        <button id="newsletter" >Get Newsletter</button>
-        
-      </div> -->
-      
-
       <section class="products" >
         <h2>Top Products</h2>
         <ul class="product-list">
         </ul>
       </section>
     </main>
+    
     <footer id="main-footer"></footer>
   </body>
 </html>
