+++ src/public/partials/header.html
@@ -2,6 +2,17 @@
   <img src="/images/noun_Tent_2517.svg" alt="tent image for logo" />
   <a href="/index.html"> Sleep<span class="highlight">Outside</span></a>
 </div>
+
+<div class="search-container">
+  
+  <form action="/action_page.php" id="search-button">
+    <input type="text" placeholder="Search..." name="search" id="inputBtn">
+    <button type="submit" id="searchBtn">
+      <img src="../images/search.png" alt="search">
+    </button>
+  </form>
+</div>
+
 <div class="cart">
   <a href="cart/index.html">
     <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100">
